update base code / make it trainable
